feat: CreateRecordFAB 리팩토링 및 RecordMenu, RecordButton 제거
- CreateRecordFAB 컴포넌트를 새로 작성하여 사용자 인터페이스 개선
- RecordMenu와 RecordButton 컴포넌트를 제거하고 CreateRecordFAB에 통합
- _layout.tsx에서 RecordMenu와 RecordButton 관련 코드 삭제
- 홈 페이지에서 CreateRecordFAB로 변경하여 기록 추가 기능을 직관적으로 제공
- useListenActiveTab 훅을 추가하여 현재 활성화된 탭 상태 관리 기능 구현
