+++ src/features/create-record-fab/ui/index.ts
@@ -1,2 +1,1 @@
-export * from './RecordButton';
-export * from './RecordMenu';
+export * from './CreateRecordFAB';
