+++ src/pages/home/ui/HomePage.tsx
@@ -1,4 +1,4 @@
-import { RecordButton } from '@features/create-record-fab';
+import { CreateRecordFAB } from '@features';
 import { HomeHeaderHama, HomeWelcome } from '@widgets';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
@@ -7,7 +7,7 @@ export const HomePage = () => {
     <SafeAreaView className="flex-1 bg-gray-50">
       <HomeHeaderHama />
       <HomeWelcome />
-      <RecordButton />
+      <CreateRecordFAB />
     </SafeAreaView>
   );
 };
